Refactor ApplicationList to presentational/container style
