+++ src/features/applicationList/ApplicationListContainer.tsx
@@ -1,31 +1,18 @@
 import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
 import { useQuery } from '@apollo/react-hooks';
 
 import ApplicationList from './ApplicationList';
-import Table, { Column, COLUMN_WIDTH } from '../../common/table/Table';
-import InternalLink from '../../common/internalLink/InternalLink';
-import ApplicationDetails from '../../common/applicationDetails/ApplicationDetails';
 import {
   BERTH_APPLICATIONS,
   BERTH_APPLICATIONSVariables as BERTH_APPLICATIONS_VARS,
 } from './__generated__/BERTH_APPLICATIONS';
-import { getBerthApplicationData, ApplicationData } from './utils';
-import { formatDate } from '../../common/utils/format';
-import Chip from '../../common/chip/Chip';
-import TableFilters from '../../common/tableFilters/TableFilters';
-import { APPLICATION_STATUS } from '../../common/utils/consonants';
+import { getBerthApplicationData } from './utils';
 import { BERTH_APPLICATIONS_QUERY } from './queries';
 import { useDeleteBerthApplication } from '../../common/mutations/deleteBerthApplication';
-import { ApplicationStatus } from '../../@types/__generated__/globalTypes';
-import Pagination from '../../common/pagination/Pagination';
 import { usePagination } from '../../common/utils/usePagination';
 import { useBackendSorting } from '../../common/utils/useBackendSorting';
 
-type ColumnType = Column<ApplicationData>;
-
 const ApplicationListContainer = () => {
-  const { t, i18n } = useTranslation();
   const [onlySwitchApps, setOnlySwitchApps] = useState<boolean>();
 
   const { cursor, pageSize, pageIndex, getPageCount, goToPage } = usePagination();
@@ -43,65 +30,6 @@ const ApplicationListContainer = () => {
 
   const [deleteDraftedApplication, { loading: isDeleting }] = useDeleteBerthApplication();
 
-  const columns: ColumnType[] = [
-    {
-      Cell: ({ cell }) => (
-        <InternalLink to={`/applications/${cell.row.original.id}`}>
-          {cell.value
-            ? t('applicationList.applicationType.switchApplication')
-            : t('applicationList.applicationType.newApplication')}
-        </InternalLink>
-      ),
-      Header: t('applicationList.tableHeaders.applicationType') || '',
-      accessor: 'isSwitch',
-      filter: 'exact',
-      disableSortBy: true,
-      width: COLUMN_WIDTH.M,
-    },
-    {
-      Cell: ({ cell }) => formatDate(cell.value, i18n.language),
-      Header: t('applicationList.tableHeaders.pvm') || '',
-      accessor: 'createdAt',
-      width: COLUMN_WIDTH.S,
-    },
-    {
-      Header: t('applicationList.tableHeaders.queue') || '',
-      accessor: 'queue',
-      disableSortBy: true,
-      width: COLUMN_WIDTH.XS,
-    },
-    {
-      Header: t('applicationList.tableHeaders.municipality') || '',
-      accessor: 'municipality',
-      disableSortBy: true,
-      width: COLUMN_WIDTH.S,
-    },
-    {
-      Cell: ({ cell: { value } }) => (
-        <Chip
-          color={APPLICATION_STATUS[value as ApplicationStatus].color}
-          label={t(APPLICATION_STATUS[value as ApplicationStatus].label)}
-        />
-      ),
-      Header: t('applicationList.tableHeaders.status') || '',
-      accessor: 'status',
-      disableSortBy: true,
-      width: COLUMN_WIDTH.M,
-    },
-    {
-      Cell: ({ cell }) =>
-        cell.value && (
-          <InternalLink to={`/harbors/${cell.value.harborId}`}>
-            {[cell.value.harborName, cell.value.pierIdentifier, cell.value.berthNum].filter(Boolean).join(' ')}
-          </InternalLink>
-        ),
-      Header: t('applicationList.tableHeaders.lease') || '',
-      accessor: 'lease',
-      disableSortBy: true,
-      width: COLUMN_WIDTH.XL,
-    },
-  ];
-
   const handleDeleteLease = async (id: string) => {
     await deleteDraftedApplication({
       variables: {
@@ -115,47 +43,19 @@ const ApplicationListContainer = () => {
   const tableData = getBerthApplicationData(data);
 
   return (
-    <ApplicationList>
-      <Table
-        data={tableData}
-        loading={loading || isDeleting}
-        columns={columns}
-        renderSubComponent={(row) => <ApplicationDetails {...row.original} handleDeleteLease={handleDeleteLease} />}
-        renderMainHeader={() => {
-          const filters = [
-            {
-              value: true,
-              label: t('applicationList.tableHeaders.switchFilter'),
-            },
-            {
-              value: false,
-              label: t('applicationList.tableHeaders.newApplicationFilter'),
-            },
-          ];
-
-          return (
-            <TableFilters
-              filters={filters}
-              activeFilters={filters.map((filter) => filter.value).filter((value) => value === onlySwitchApps)}
-              handleSetFilter={(filter) => {
-                goToPage(0);
-                setOnlySwitchApps(filter);
-              }}
-            />
-          );
-        }}
-        renderTableToolsBottom={() => (
-          <Pagination
-            forcePage={pageIndex}
-            pageCount={getPageCount(data?.berthApplications?.count)}
-            onPageChange={({ selected }) => goToPage(selected)}
-          />
-        )}
-        renderEmptyStateRow={() => t('common.notification.noData.description')}
-        onSortedColChange={handleSortedColChange({ createdAt: 'createdAt' })}
-        canSelectRows
-      />
-    </ApplicationList>
+    <ApplicationList
+      data={data}
+      getPageCount={getPageCount}
+      goToPage={goToPage}
+      handleDeleteLease={handleDeleteLease}
+      onSortedColChange={handleSortedColChange({ createdAt: 'createdAt' })}
+      isDeleting={isDeleting}
+      loading={loading}
+      onlySwitchApps={onlySwitchApps}
+      pageIndex={pageIndex}
+      setOnlySwitchApps={setOnlySwitchApps}
+      tableData={tableData}
+    />
   );
 };
